pyproject.toml
[build-system]
requires = ["setuptools>=59.6.0"]
build-backend = "setuptools.build_meta"

[project]
# changes for these values should be reflected in docs/source/conf.py as well
name = "neurosnap"
version = "2025.09.09"
authors = [
  {name="Keaun Amani"},
]
description = "Collection of useful bioinformatic functions and tools for various computational biology pipelines."
license = {file = "LICENSE"}
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Development Status :: 3 - Alpha",
]
requires-python = ">=3.8"
dependencies = [
  "requests>2",
  "numpy>1.0.0",
  "pandas",
  "biopython",
  "rdkit>2023",
  "tqdm",
  "scipy",
  "matplotlib",
  "tabulate",
  "requests_toolbelt",
]

[project.optional-dependencies]
dev = [
  "build",
  "pytest",
  "twine==5.1.1",
  "ruff==0.7.0",
  "sphinx==8.1.3",
  "sphinx-rtd-theme==3.0.1",
  "sphinx-autodoc-typehints==2.5.0",
]

clusterprot = [
  "scikit-learn",
  "umap-learn",
  "plotly"
]

kluster = [
  "scikit-learn",
  "umap-learn",
  "plotly"
]

[project.urls]
Homepage = "https://github.com/NeurosnapInc/neurosnap"
Issues = "https://github.com/NeurosnapInc/neurosnap/issues"

[tool.ruff]
line-length = 150
indent-width = 2

[tool.pytest.ini_options]
addopts = "-ra"
markers = [
  "slow: may take >10s",
  "integration: crosses module boundaries / uses real IO",
  "mmseqs2: requires ColabFold MMseqs2 API/network access",
]